% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{clause}
\alias{clause}
\title{Create a clause}
\usage{
clause(name, ...)
}
\arguments{
\item{name}{a length 1 character, the name of the clause.}

\item{...}{any number of valid arguments. I.e. int, atom, variable or char.}
}
\description{
Create a clause
}
