% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{head.logician_iterator}
\alias{head.logician_iterator}
\title{Return the first n results from an iterator}
\usage{
\method{head}{logician_iterator}(x, n = 6, ...)
}
\arguments{
\item{x}{an object}

\item{n}{an length 1 integer >= 1.
sometimes that can result in an infinite loop}

\item{...}{}
}
\value{
For each solution a list of substitions.
}
\description{
Return the first n results from an iterator
}
\examples{
database <- logician_database(
  connected(berlin, hamburg),
  connected(hamburg, chicago),
  connected(chicago, london),
  connected(aachen, berlin),
  connected(chicago, portland),
  connected(portland, munich),
  path(A, B) := connected(A, B),
  path(A, B) := connected(A, Z) && path(Z, B)
)
iter <- logician_query(database, path(berlin, X))
head(iter, 4) # get the first 4 results if they exists

}
