% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{logician_query}
\alias{logician_query}
\alias{logician_query_}
\alias{logician_query_.logician_clause}
\alias{logician_query_.logician_r_expr}
\title{Query a database}
\usage{
logician_query(database, question)

logician_query_(database, question)

\method{logician_query_}{logician_clause}(database, question)

\method{logician_query_}{logician_r_expr}(database, question)
}
\arguments{
\item{database}{a list of rules and facts}

\item{question}{a clause of type \code{logician_clause} or an R expression
of type \code{logician_r_expr}}
}
\description{
Query a database
}
\examples{
database <- logician_database(
  connected(berlin, hamburg),
  connected(hamburg, chicago),
  connected(chicago, london),
  connected(aachen, berlin),
  connected(chicago, portland),
  connected(portland, munich),
  path(A, B) := connected(A, B),
  path(A, B) := connected(A, Z) && path(Z, B)
)
iter <- logician_query(database, path(berlin, hamburg))
iter$next_value()

iter <- logician_query(database, path(berlin, X))
iter$next_value()
iter$next_value()
iter$next_value()
}
