% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{r_expr}
\alias{r_expr}
\title{Create an r expression}
\usage{
r_expr(code)
}
\arguments{
\item{code}{valid R code that returns a length 1 logical. Any symbols with
uppercase first letter will be treated as internal variables. You can only
refer to names from the globalenv here.}
}
\description{
Create an r expression
}
