% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{rule}
\alias{rule}
\title{Create a rule}
\usage{
rule(head, body)
}
\arguments{
\item{head}{a head is always a clause}

\item{body}{a body is a list of clauses or r expressions.}
}
\description{
Create a rule
}
